requirements.txt
# Async HTTP client
aiohttp==3.9.1
aiofiles==23.2.1

# Core dependencies  
certifi==2024.12.14
charset-normalizer==3.4.1
idna==3.10

# YAML configuration
PyYAML==6.0.1

# Optional: for better console output
tqdm==4.66.1
colorama==0.4.6

# For type hints (optional but recommended)
typing-extensions==4.9.0
